+++ docs/changes.rst
@@ -7,6 +7,8 @@ Version 0.1.7
 To be released. Still alpha version.
 
  - Added :attr:`Image.mimetype <wand.image.Image.mimetype>` property.
+ - Added :meth:`Image.crop() <wand.image.Image.crop>` method for in-place
+   crop.
 
 Version 0.1.6
 -------------